chore/destination_elder: extract out ADataHandler
